Don't cast the request of an ioctl to an int. Bad juju. This plus ross's earlier work fixes PR xsrc/7061 from Brett Slager.

+++ xc/programs/Xserver/os/access.c
@@ -666,7 +666,7 @@ DefineSelf (fd)
 #endif
     ifc.ifc_len = sizeof (buf);
     ifc.ifc_buf = buf;
-    if (ifioctl (fd, (int) SIOCGIFCONF, (pointer) &ifc) < 0)
+    if (ifioctl (fd, SIOCGIFCONF, (pointer) &ifc) < 0)
         Error ("Getting interface configuration (4)");
 
 #ifdef ISC
